+++ jest.config.ts
@@ -3,5 +3,15 @@ export default {
   coverageDirectory: "coverage",
   coverageProvider: "v8",
   testEnvironment: "node",
-  modulePathIgnorePatterns: ["/node_modules/", "/utils/"],
+  modulePathIgnorePatterns: ["/node_modules/"],
+  testMatch: ["**/__tests__/**/*.spec.ts", "**/__tests__/**/*.test.ts"],
+  collectCoverageFrom: [
+    "./packages/**/*.ts",
+    "!./packages/**/*.js",
+    "!./packages/**/*.cjs",
+    "!./packages/**/*.d.ts",
+    "!./packages/*/__tests__/**",
+    "!./packages/**/index.ts",
+    "!./packages/**/types.ts",
+  ],
 };