+++ packages/core/src/index.ts
@@ -1,3 +1,9 @@
 export { default as Benzene } from "./core";
-export type { ExtractExtraType, GraphQLParams, ValueOrPromise } from "./types";
-export { isAsyncIterator } from "./utils";
+export type {
+  CompiledQuery,
+  CompileQuery,
+  ExtractExtraType,
+  GraphQLParams,
+  ValueOrPromise,
+} from "./types";
+export { isAsyncIterator, makeCompileQuery } from "./utils";