+++ packages/jit/rollup.config.js
@@ -2,5 +2,5 @@ import merge from "deepmerge";
 import config from "../../rollup.config";
 
 export default merge(config, {
-  external: ["@benzene/core", "@hoangvvo/graphql-jit", "tiny-lru"],
-})
+  external: ["@benzene/core", "@hoangvvo/graphql-jit"],
+});