+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @benzene/core
 
+## 0.6.0-next.1
+
+### Patch Changes
+
+- Build files at root
+
+## 0.6.0-next.0
+
+### Minor Changes
+
+- 055eda0: allow custom compileQuery implementation
+
 ## 0.5.0
 
 ### Minor Changes